+++ src/fhirpath/__init__.py
@@ -5,7 +5,10 @@
 def get_version():
     """ """
     import os
-    with open(os.path.join(os.path.dirname(os.path.abspath(__file__)), "version.py"), "r") as fp:
+
+    with open(
+        os.path.join(os.path.dirname(os.path.abspath(__file__)), "version.py"), "r"
+    ) as fp:
         for line in fp:
             ln = line.strip()
             if not ln: