+++ src/fhirpath/fql/__init__.py
@@ -4,12 +4,14 @@ from .expressions import G_
 from .expressions import T_
 from .expressions import V_
 from .expressions import and_
+from .expressions import eb_
 from .expressions import exists_
 from .expressions import in_
 from .expressions import not_
 from .expressions import not_exists_
 from .expressions import not_in_
 from .expressions import or_
+from .expressions import sa_
 from .expressions import sort_
 from .types import ElementPath
 
@@ -28,5 +30,7 @@ __all__ = [
     "not_in_",
     "or_",
     "sort_",
+    "sa_",
+    "eb_",
     "ElementPath"
 ]